# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seccloud API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/intake/raw-events": {
      "post": {
        "summary": "Intake Raw Events",
        "operationId": "intake_raw_events_api_intake_raw_events_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stream/advance": {
      "post": {
        "summary": "Stream Advance",
        "operationId": "stream_advance_api_stream_advance_post",
        "parameters": [
          {
            "name": "batch_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 5,
              "title": "Batch Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamAdvance"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/detections": {
      "get": {
        "summary": "Detections",
        "operationId": "detections_api_detections_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/detections/{detection_id}": {
      "get": {
        "summary": "Detection Detail",
        "operationId": "detection_detail_api_detections__detection_id__get",
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Detection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/detections/{detection_id}/acknowledge": {
      "post": {
        "summary": "Detection Acknowledge",
        "operationId": "detection_acknowledge_api_detections__detection_id__acknowledge_post",
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Detection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events/{event_id}": {
      "get": {
        "summary": "Event Detail",
        "operationId": "event_detail_api_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "verb": {
            "type": "string",
            "title": "Verb"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": ["source", "verb", "category"],
        "title": "Action"
      },
      "Detection": {
        "properties": {
          "detection_id": {
            "type": "string",
            "title": "Detection Id"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "feature_attributions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Feature Attributions"
          },
          "event_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Event Ids"
          },
          "related_entity_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Entity Ids"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidencePointer"
            },
            "type": "array",
            "title": "Evidence"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "model_rationale": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectionModelRationale"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "detection_id",
          "scenario",
          "title",
          "score",
          "confidence",
          "severity",
          "reasons",
          "feature_attributions",
          "event_ids",
          "related_entity_ids",
          "evidence",
          "model_version",
          "status"
        ],
        "title": "Detection"
      },
...
      "DetectionDetail": {
        "properties": {
          "detection": {
            "$ref": "#/components/schemas/Detection"
          },
          "peer_comparison": {
            "$ref": "#/components/schemas/DetectionPeerComparison"
          },
          "evidence_bundle": {
            "$ref": "#/components/schemas/DetectionEvidenceBundle"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "detection",
          "peer_comparison",
          "evidence_bundle",
          "events"
        ],
        "title": "DetectionDetail"
      },
      "DetectionEvidenceBundle": {
        "properties": {
          "evidence_items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceBundleItem"
            },
            "type": "array",
            "title": "Evidence Items"
          }
        },
        "type": "object",
        "required": ["evidence_items"],
        "title": "DetectionEvidenceBundle"
      },
      "DetectionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": ["items", "page"],
        "title": "DetectionList"
      },
      "DetectionModelRationale": {
        "properties": {
          "scoring_mode": {
            "type": "string",
            "title": "Scoring Mode"
          },
          "policy_scope": {
            "type": "string",
            "title": "Policy Scope"
          },
          "detection_threshold": {
            "type": "number",
            "title": "Detection Threshold"
          },
          "high_severity_threshold": {
            "type": "number",
            "title": "High Severity Threshold"
          },
          "model_score": {
            "type": "number",
            "title": "Model Score"
          },
          "score_margin": {
            "type": "number",
            "title": "Score Margin"
          },
          "calibration_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Source"
          },
          "calibration_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Reason"
          }
        },
        "type": "object",
        "required": [
          "scoring_mode",
          "policy_scope",
          "detection_threshold",
          "high_severity_threshold",
          "model_score",
          "score_margin"
        ],
        "title": "DetectionModelRationale"
      },
      "DetectionPeerComparison": {
        "properties": {
          "principal_id": {
            "type": "string",
            "title": "Principal Id"
          },
          "peer_group": {
            "type": "string",
            "title": "Peer Group"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "principal_role": {
            "type": "string",
            "title": "Principal Role"
          },
          "principal_location": {
            "type": "string",
            "title": "Principal Location"
          },
          "principal_privilege_level": {
            "type": "string",
            "title": "Principal Privilege Level"
          },
          "principal_total_events": {
            "type": "integer",
            "title": "Principal Total Events"
          },
          "principal_prior_event_count": {
            "type": "integer",
            "title": "Principal Prior Event Count"
          },
          "principal_prior_action_count": {
            "type": "integer",
            "title": "Principal Prior Action Count"
          },
          "principal_prior_resource_access_count": {
            "type": "integer",
            "title": "Principal Prior Resource Access Count"
          },
          "peer_group_resource_access_count": {
            "type": "integer",
            "title": "Peer Group Resource Access Count"
          },
          "peer_group_principal_count": {
            "type": "integer",
            "title": "Peer Group Principal Count"
          },
          "peer_group_resource_principal_count": {
            "type": "integer",
            "title": "Peer Group Resource Principal Count"
          },
          "department_peer_count": {
            "type": "integer",
            "title": "Department Peer Count"
          },
          "manager_peer_count": {
            "type": "integer",
            "title": "Manager Peer Count"
          },
          "group_peer_count": {
            "type": "integer",
            "title": "Group Peer Count"
          },
          "geo_seen_before": {
            "type": "boolean",
            "title": "Geo Seen Before"
          },
          "detection_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detection Reasons"
          }
        },
        "type": "object",
        "required": [
          "principal_id",
          "peer_group",
          "resource_id",
          "principal_role",
          "principal_location",
          "principal_privilege_level",
          "principal_total_events",
          "principal_prior_event_count",
          "principal_prior_action_count",
          "principal_prior_resource_access_count",
          "peer_group_resource_access_count",
          "peer_group_principal_count",
          "peer_group_resource_principal_count",
          "department_peer_count",
          "manager_peer_count",
          "group_peer_count",
          "geo_seen_before",
          "detection_reasons"
        ],
        "title": "DetectionPeerComparison"
      },
      "Event": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_key": {
            "type": "string",
            "title": "Event Key"
          },
          "integration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integration Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_event_id": {
            "type": "string",
            "title": "Source Event Id"
          },
          "principal": {
            "$ref": "#/components/schemas/Principal"
          },
          "resource": {
            "$ref": "#/components/schemas/Resource"
          },
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          },
          "environment": {
            "additionalProperties": true,
            "type": "object",
            "title": "Environment"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "evidence": {
            "$ref": "#/components/schemas/EvidencePointer"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_key",
          "source",
          "source_event_id",
          "principal",
          "resource",
          "action",
          "observed_at",
          "environment",
          "attributes",
          "evidence"
        ],
        "title": "Event"
      },
...
      "EventList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": ["items", "page"],
        "title": "EventList"
      },
...
      "EvidencePointer": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "raw_event_id": {
            "type": "string",
            "title": "Raw Event Id"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          }
        },
        "type": "object",
        "required": ["source", "object_key", "raw_event_id", "observed_at"],
        "title": "EvidencePointer"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntakeAccepted": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "integration_id": {
            "type": "string",
            "title": "Integration Id"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "manifest_key": {
            "type": "string",
            "title": "Manifest Key"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate"
          },
          "queue_path": {
            "type": "string",
            "title": "Queue Path"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "tenant_id",
          "source",
          "integration_id",
          "record_count",
          "manifest_key",
          "object_key",
          "idempotency_key",
          "duplicate",
          "queue_path"
        ],
        "title": "IntakeAccepted"
      },
...
      "Pagination": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "returned": {
            "type": "integer",
            "title": "Returned"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": ["limit", "offset", "returned", "has_more"],
        "title": "Pagination"
      },
...
      "StreamAdvance": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cursor": {
            "type": "integer",
            "title": "Cursor"
          },
          "total_source_events": {
            "type": "integer",
            "title": "Total Source Events"
          },
          "complete": {
            "type": "boolean",
            "title": "Complete"
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size"
          },
          "accepted_batches": {
            "type": "integer",
            "title": "Accepted Batches"
          },
          "accepted_records": {
            "type": "integer",
            "title": "Accepted Records"
          },
          "pending_batch_count": {
            "type": "integer",
            "title": "Pending Batch Count"
          }
        },
        "type": "object",
        "required": [
          "status",
          "cursor",
          "total_source_events",
          "complete",
          "batch_size",
          "accepted_batches",
          "accepted_records",
          "pending_batch_count"
        ],
        "title": "StreamAdvance"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}